requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0
openpyxl==3.1.2

# Calendar
icalendar==5.0.11

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP client (for Amion scraping)
httpx==0.26.0
playwright==1.41.0

# LLM integration
openai==1.12.0

# Environment
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2

# Production
gunicorn==21.2.0

# Development/Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
ruff==0.1.14
